Remove unused files, removed demo page, add event for Gimg
