+++ src/App.vue
@@ -6,14 +6,11 @@
 			<router-link to="/slider">Slider demo</router-link>
 			<router-link to="/menu-demo">Menu</router-link>
 			<router-link to="/pulse-demo">Pulse</router-link>
-			<router-link to="/card">Card</router-link>
 			<router-link to="/overlay-demo">Overlay</router-link>
 			<router-link to="/dialog-demo">Dialog Demo</router-link>
       <router-link to="/tool-tip-demo">Tooltip demo</router-link>
-			<router-link to="/button">Button</router-link>
 			<router-link to="/input">Input</router-link>
 			<router-link to="/transition-demo">Transition</router-link>
-			<router-link to="/chip">Chip</router-link>
 			<router-link to="/list-demo">List</router-link>
 			<router-link to="/dnd-dialog-demo">Dnd Dialog Demo</router-link>
 			<router-link to="/selection-control">SelectIonControl</router-link>
