+++ src/plugin/router.js
@@ -4,8 +4,6 @@ import KeyboardDemo from '@/view/KeyboardDemo';
 import Groupable from '@/view/Groupable';
 import MenuDemo from '@/view/MenuDemo';
 import SliderDemo from '../view/SliderDemo';
-import CardDemo from '../view/CardDemo';
-import ButtonDemo from '../view/ButtonDemo';
 import OverlayDemo from "../view/OverlayDemo";
 import SelectIonControl from '../view/SelectIonControl';
 import DialogDemo from '../view/DialogDemo';
@@ -13,7 +11,6 @@ import PulseDemo from '../view/PulseDemo';
 import ToolTipDemo from '../view/ToolTipDemo';
 import InputDemo from "../view/InputDemo";
 import TransitionDemo from '../view/TransitionDemo';
-import ChipDemo from '../view/ChipDemo';
 import ListDemo from '../view/ListDemo';
 import List2Demo from "../view/List2Demo";
 import DndDialogDemo from '../view/DndDialogDemo';
